Text_Summarizer_with_Gemini/requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
nltk==3.8.1 